requirements.txt
streamlit
pandas
plotly
simple-salesforce
python-dotenv
pytz
numpy
matplotlib
scikit-learn
dspy
toml
black
flake8
passlib
bcrypt
fuzzywuzzy
python-Levenshtein